Put bed download into helper file
